constoller, service and model lost weigth
