+++ app/Http/Controllers/System/AppsController.php
@@ -3,13 +3,15 @@
 namespace App\Http\Controllers\System;
 
 use App\Services\System\AppsService;
+use App\Models\System\App;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
 class AppsController extends Controller
 {
-    public function __construct(AppsService $service)
+    public function __construct(App $model, AppsService $service)
     {
+        $this->model = $model;
         $this->service = $service;
     }
 
@@ -17,35 +19,5 @@ class AppsController extends Controller
     {
         $menu = $this->service->getMenu();
         return response()->json($menu);
-    }
-    
-    public function index()
-    {
-        $index = $this->service->appList();
-        return response()->json($index);
-    }
-    
-    public function create(Request $request)
-    {
-        $create = $this->service->createApps($resquest);
-        return response()->json($create);
-    }
-
-    public function edit($id)
-    {
-        $edit = $this->service->getApp($id);
-        return response()->json($edit);
-    }
-
-    public function update(Request $request, $id)
-    {
-        $update = $this->service->updateApp($params, $id);
-        return response()->json($update);
-    }
-
-    public function destroy($id)
-    {
-        $destroy = $this->service->deleteApp($id);
-        return response()->json($destroy);
     }
 }
